# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Shopee Affiliate Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/top-products:
    get:
      operationId: getTopProducts
      tags: [stats]
      summary: Get top 10 most clicked products
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /stats/clicks-chart:
    get:
      operationId: getClicksChart
      tags: [stats]
      summary: Get daily click data for chart
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Click chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClickChartPoint"
...
components:
  schemas:
...
    TopProduct:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
        clickCount:
          type: integer
        viewCount:
          type: integer
        commission:
          type: integer
          nullable: true
        commissionRate:
          type: number
          nullable: true
        price:
          type: integer
        status:
          type: string
      required: [id, name, slug, imageUrl, clickCount, viewCount, price, status]
    ClickChartPoint:
      type: object
      properties:
        date:
          type: string
        clicks:
          type: integer
        views:
          type: integer
      required: [date, clicks, views]
...